Refactor methods for fetching proj from URL into mixin
